refactor: extract DeveloperToolsScreen logic into custom hooks
Decompose 1080-line DeveloperToolsScreen into smaller, reusable hooks:
- useDiagnostics: System health checks (database, Sentry, error logs)
- useNotificationTesting: Notification testing and diagnostics
- useDatabaseOperations: Database reset operations
- useSentryTesting: Sentry integration testing
- useErrorLogManagement: Error log viewing and management

Benefits:
- Reduces DeveloperToolsScreen from 1080 to 572 lines (47% reduction)
- Improves code organization and separation of concerns
- Enhances testability by isolating logic into focused hooks
- Makes developer tools easier to maintain and extend

All tests pass (2073/2073). Zero functionality changes.
